Moved unimplemented api definitions #252
Definitions now live in `tentative-lock-keeper-api` repo
