+++ lock-keeper-client/src/api/retrieve_audit_events.rs
@@ -1,6 +1,6 @@
-use crate::{api::AuditEventOptions, LockKeeperClient, LockKeeperClientError};
+use crate::{LockKeeperClient, LockKeeperClientError};
 use lock_keeper::{
-    audit_event::{AuditEvent, EventType},
+    audit_event::{AuditEvent, AuditEventOptions, EventType},
     channel::ClientChannel,
     types::retrieve_audit_events::{client, server},
 };
